Move deprecation test helper into fixtures folder.
